requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
web3==6.15.1
python-dateutil==2.8.2
